requirements.txt
ruamel.yaml==0.18.*
pytest==7.*
